src/pmdas/amdgpu: move fake library routines to a separate source file
Move fake routines to (new) fake.c and cleanup drm.c.
